backend/pyproject.toml
[project]
name = "blakjaks-platform"
version = "0.1.0"
description = "BlakJaks platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0",
    "httpx[http2]>=0.26.0",
    "cryptography>=41.0.0",
    "web3>=6.15.0",
    "google-cloud-kms>=2.21.0",
    "google-cloud-vision>=3.5.0",
    "google-cloud-storage>=2.14.0",
    "Pillow>=10.2.0",
    # Redis
    "redis[asyncio]>=5.0.0",
    # Celery
    "celery>=5.3.0",
    # Auth (Argon2id — replaces bcrypt)
    "argon2-cffi>=21.3.0",
    "passlib[argon2]>=1.7.4",
    # Sentry
    "sentry-sdk[fastapi]>=1.40.0",
    # Prometheus
    "prometheus-fastapi-instrumentator>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.20.0",
    "fakeredis[aioredis]>=2.20.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
